Rename packages: OpenApi → OpenApi.Microsoft, Swashbuckle → OpenApi.Swashbuckle
The package previously called Kalicz.StrongTypes.OpenApi specifically targets
Microsoft's AddOpenApi pipeline, so its name implied genericity it didn't have.
Renaming to make the parallel structure obvious to consumers picking between
the two pipelines.

Also moves StrongTypes.SwaggerApi → StrongTypes.OpenApi.TestApi.Swashbuckle
and consolidates both packages + the Swashbuckle test app under a single
'OpenApi packages/' solution folder.

Mechanical rename only — no behaviour change. Solution builds clean.
