+++ StrongTypes.slnx
@@ -2,7 +2,6 @@
   <Folder Name="/API Tests/">
     <Project Path="src/StrongTypes.Api.IntegrationTests/StrongTypes.Api.IntegrationTests.csproj" />
     <Project Path="src/StrongTypes.Api/StrongTypes.Api.csproj" />
-    <Project Path="src/StrongTypes.SwaggerApi/StrongTypes.SwaggerApi.csproj" />
   </Folder>
   <Folder Name="/EF Core package/">
     <Project Path="src/StrongTypes.EfCore/StrongTypes.EfCore.csproj" />
@@ -10,11 +9,10 @@
   <Folder Name="/FsCheck package/">
     <Project Path="src/StrongTypes.FsCheck/StrongTypes.FsCheck.csproj" />
   </Folder>
-  <Folder Name="/OpenApi package/">
-    <Project Path="src/StrongTypes.OpenApi/StrongTypes.OpenApi.csproj" />
-  </Folder>
-  <Folder Name="/Swashbuckle package/">
-    <Project Path="src/StrongTypes.Swashbuckle/StrongTypes.Swashbuckle.csproj" />
+  <Folder Name="/OpenApi packages/">
+    <Project Path="src/StrongTypes.OpenApi.Microsoft/StrongTypes.OpenApi.Microsoft.csproj" />
+    <Project Path="src/StrongTypes.OpenApi.Swashbuckle/StrongTypes.OpenApi.Swashbuckle.csproj" />
+    <Project Path="src/StrongTypes.OpenApi.TestApi.Swashbuckle/StrongTypes.OpenApi.TestApi.Swashbuckle.csproj" />
   </Folder>
   <Folder Name="/Main Package/">
     <Project Path="src/StrongTypes.Analyzers/StrongTypes.Analyzers.csproj" />
