+++ src/StrongTypes.OpenApi.Microsoft/readme.md
