+++ src/StrongTypes.OpenApi.Swashbuckle/StrongTypes.OpenApi.Swashbuckle.csproj
@@ -1,0 +1,35 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFramework>net10.0</TargetFramework>
+    <LangVersion>14.0</LangVersion>
+    <Nullable>enable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
+    <NoWarn>CS1591</NoWarn>
+
+    <Version>0.0.0-dev</Version>
+    <PackageId>Kalicz.StrongTypes.OpenApi.Swashbuckle</PackageId>
+    <RootNamespace>StrongTypes.OpenApi.Swashbuckle</RootNamespace>
+    <Authors>KaliCZ</Authors>
+    <Copyright>Copyright © 2026 KaliCZ</Copyright>
+    <Description>Schema filters that wire Kalicz.StrongTypes into Swashbuckle.AspNetCore (AddSwaggerGen). Makes the generated Swagger document describe each strong-type wrapper with the JSON its converter actually reads and writes (NonEmptyString as string with minLength:1, Positive&lt;T&gt;/NonNegative&lt;T&gt;/Negative&lt;T&gt;/NonPositive&lt;T&gt; as the underlying primitive with the appropriate minimum/maximum, NonEmptyEnumerable&lt;T&gt; as array with minItems:1, Maybe&lt;T&gt; as the wire-format object wrapper). For Microsoft.AspNetCore.OpenApi (AddOpenApi), use Kalicz.StrongTypes.OpenApi.Microsoft instead.</Description>
+    <PackageTags>StrongTypes, Swashbuckle, Swagger, OpenApi, AspNetCore, SchemaFilter, NonEmptyString, NonEmptyEnumerable, Maybe</PackageTags>
+    <PackageLicenseExpression>MIT</PackageLicenseExpression>
+    <PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
+    <PackageProjectUrl>https://github.com/KaliCZ/StrongTypes</PackageProjectUrl>
+    <RepositoryType>git</RepositoryType>
+    <RepositoryUrl>https://github.com/KaliCZ/StrongTypes.git</RepositoryUrl>
+    <PackageReadmeFile>readme.md</PackageReadmeFile>
+    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+    <GenerateDocumentationFile>true</GenerateDocumentationFile>
+  </PropertyGroup>
+  <ItemGroup>
+    <PackageReference Include="Swashbuckle.AspNetCore.SwaggerGen" Version="10.1.7" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\StrongTypes\StrongTypes.csproj" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="readme.md" Pack="true" PackagePath="\" />
+  </ItemGroup>
+</Project>
