+++ src/StrongTypes.OpenApi.TestApi.Swashbuckle/StrongTypes.OpenApi.TestApi.Swashbuckle.csproj
@@ -1,0 +1,20 @@
+<Project Sdk="Microsoft.NET.Sdk.Web">
+  <PropertyGroup>
+    <TargetFramework>net10.0</TargetFramework>
+    <LangVersion>14.0</LangVersion>
+    <Nullable>enable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
+    <NoWarn>CS1591</NoWarn>
+    <IsPackable>false</IsPackable>
+    <RootNamespace>StrongTypes.OpenApi.TestApi.Swashbuckle</RootNamespace>
+    <AssemblyName>StrongTypes.OpenApi.TestApi.Swashbuckle</AssemblyName>
+  </PropertyGroup>
+  <ItemGroup>
+    <PackageReference Include="Swashbuckle.AspNetCore" Version="10.1.7" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\StrongTypes\StrongTypes.csproj" />
+    <ProjectReference Include="..\StrongTypes.OpenApi.Swashbuckle\StrongTypes.OpenApi.Swashbuckle.csproj" />
+  </ItemGroup>
+</Project>
