+++ src/StrongTypes.OpenApi.TestApi.Swashbuckle/appsettings.json
