--- src/StrongTypes.OpenApi/StrongTypes.OpenApi.csproj
@@ -1,35 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net10.0</TargetFramework>
-    <LangVersion>14.0</LangVersion>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-    <NoWarn>CS1591</NoWarn>
-
-    <Version>0.0.0-dev</Version>
-    <PackageId>Kalicz.StrongTypes.OpenApi</PackageId>
-    <Authors>KaliCZ</Authors>
-    <Copyright>Copyright © 2026 KaliCZ</Copyright>
-    <Description>OpenAPI schema transformers for Kalicz.StrongTypes. Makes Microsoft.AspNetCore.OpenApi produce schemas that match the JSON the strong-type converters actually read and write (NonEmptyString as string with minLength:1, Positive&lt;T&gt;/NonNegative&lt;T&gt;/Negative&lt;T&gt;/NonPositive&lt;T&gt; as the underlying primitive with the appropriate minimum/maximum, NonEmptyEnumerable&lt;T&gt; as array with minItems:1, Maybe&lt;T&gt; as the wire-format object wrapper).</Description>
-    <PackageTags>StrongTypes, OpenApi, Swagger, AspNetCore, SchemaTransformer, NonEmptyString, NonEmptyEnumerable, Maybe</PackageTags>
-    <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
-    <PackageProjectUrl>https://github.com/KaliCZ/StrongTypes</PackageProjectUrl>
-    <RepositoryType>git</RepositoryType>
-    <RepositoryUrl>https://github.com/KaliCZ/StrongTypes.git</RepositoryUrl>
-    <PackageReadmeFile>readme.md</PackageReadmeFile>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <GenerateDocumentationFile>true</GenerateDocumentationFile>
-  </PropertyGroup>
-  <ItemGroup>
-    <FrameworkReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="10.0.0" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\StrongTypes\StrongTypes.csproj" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="readme.md" Pack="true" PackagePath="\" />
-  </ItemGroup>
-</Project>
